@@ -1,11 +1,8 @@
-import Vue from "vue"
-import Vuex from "vuex"
+import { createStore } from "vuex"
 import Auth from "./modules/auth"
 import persistUserSession from "@/store/plugins/persistUserSession"
 
-Vue.use(Vuex)
-
-export default new Vuex.Store({
+export default createStore({
   plugins: [persistUserSession],
   state: {
     notifications: [],
